Re-organized a bunch of code. Mouse and keyboard actions work. Woot.
